Move arg parsing to a mod file
